fix(wsClient): Reorganize files
Remove index.ts files in favor of explicit names
