+++ packages/client/rollup.config.ts
@@ -8,7 +8,7 @@ export const input = [
   'src/links/httpLink.ts',
   'src/links/loggerLink.ts',
   'src/links/splitLink.ts',
-  'src/links/wsLink/index.ts',
+  'src/links/wsLink/wsLink.ts',
   'src/unstable-internals.ts',
 ];
