+++ packages/client/src/links.ts
@@ -6,7 +6,7 @@ export * from './links/httpBatchStreamLink';
 export * from './links/httpLink';
 export * from './links/loggerLink';
 export * from './links/splitLink';
-export * from './links/wsLink';
+export * from './links/wsLink/wsLink';
 export * from './links/httpSubscriptionLink';
 export * from './links/retryLink';
